updated unit tests
